+++ rh-admin-utils.php
@@ -2,7 +2,7 @@
 
 /**
  * Plugin Name: RH Admin Utilities
- * Version: 1.8.1
+ * Version: 1.8.2
  * Author: Rasso Hilber
  * Description: Admin Utilities for WordPress. Removes plugin ads, adds custom buttons to the admin bar (publish, clear cache), allows editors to add users (except administrators), disables comments. Provides filters to adjust functionality.
  * Author URI: https://rassohilber.com
@@ -52,4 +52,5 @@ RemoveAds::getInstance();
 AdminDashboard::init();
 WpCliCommands::init();
 ACFCodeField::init();
+ACFTextField::init();
 Misc::getInstance();